Move PrepareCommentView into "comment" module.
This combines all urls that begin with "comment" into the same module, which
will make them easier to move.
